Add external-url to CSS post
According to the Octopress documentation, this will link the title of
the post to the external-url specified:

  http://octopress.org/docs/blogging/linklog/

However, the effect this has is dictated by the theme used. We are using
a theme called Greyshade, which has not implemented this feature.
Perhaps, one day, this will actually work. An issue was filed at:

  https://github.com/shashankmehta/greyshade/issues/30

Change-Id: I6f91fcd4535727557937d3a0689281c9a9667a18
Reviewed-on: https://gerrit.causes.com/22450
Tested-by: Joe Lencioni <joe.lencioni@causes.com>
Reviewed-by: Jeremy Dunck <jeremy.dunck@causes.com>