Make author info show on subpages, not just index
I have only a vague understanding of how all this Jekyll stuff works,
but the `post.author` is only set on the index page, and on subpages, it
is `page.author` which is set. Oddly, even on subpages, `post.title` is
set.

In any case, this change makes the author info appear everywhere.

It pains me, but those are hard tabs, out of respect for the local style
in the file.

Change-Id: Ieac8fce01aead200cf1946c9ff5c3bc20588bcf2
Reviewed-on: https://gerrit.causes.com/26309
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Joe Lencioni <joe.lencioni@causes.com>